flask-app/requirements.txt
cryptography==43.0.1 ; python_version >= "3.10" and python_version < "4.0"
flask==3.0.3 ; python_version >= "3.10" and python_version < "4.0"
pydantic==2.9.2 ; python_version >= "3.10" and python_version < "4.0"
PyMySQL==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==2.0.30 ; python_version >= "3.10" and python_version < "4.0"


